+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Respect return type for unbound actions/functions for message target resolution
+- Convert transient and absolute message targets correctly
+- Access prepared EDMX files if existing, otherwise compile like before
+- Register build plugin to prepare V2 metadata (enabled via adapter option `build: true`)
 
 ## Version 1.11.9 - 2023-11-20